+++ internal/provider/types/spec_renderer.go
@@ -7,10 +7,11 @@ import (
 )
 
 type SpecRenderer struct {
-	HideDeprecated types.Bool `tfsdk:"hide_deprecated"`
-	HideInternal   types.Bool `tfsdk:"hide_internal"`
-	InfiniteScroll types.Bool `tfsdk:"infinite_scroll"`
-	ShowSchemas    types.Bool `tfsdk:"show_schemas"`
-	TryItInsomnia  types.Bool `tfsdk:"try_it_insomnia"`
-	TryItUI        types.Bool `tfsdk:"try_it_ui"`
+	AllowCustomServerUrls types.Bool `tfsdk:"allow_custom_server_urls"`
+	HideDeprecated        types.Bool `tfsdk:"hide_deprecated"`
+	HideInternal          types.Bool `tfsdk:"hide_internal"`
+	InfiniteScroll        types.Bool `tfsdk:"infinite_scroll"`
+	ShowSchemas           types.Bool `tfsdk:"show_schemas"`
+	TryItInsomnia         types.Bool `tfsdk:"try_it_insomnia"`
+	TryItUI               types.Bool `tfsdk:"try_it_ui"`
 }